+++ pytest_lambda/__init__.py
@@ -1,6 +1,6 @@
 """Define pytest fixtures using lambda functions"""
 
-__version__ = '1.3.0'
+__version__ = '2.1.0'
 
 from .fixtures import *
 from .util import *