+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
+
+
+## [2.1.0] — 2022-07-17
 ### Changed
  - Preserve declared order of arguments with `wrap_fixture` (decorated method's first, then wrapped fixture's, then `request`)
  - DOC: add destructuring examples to README